requirements.txt
# Core Dependencies
langgraph>=0.2.16
langchain>=0.2.14
langchain-openai>=0.1.23
langchain-community>=0.2.12

# AI/ML Libraries
openai>=1.40.0
anthropic>=0.34.0

# Data Processing
pandas==2.2.2
numpy==1.26.4
pydantic==2.8.2

# Security & Threat Intelligence
requests==2.32.3
python-dotenv==1.0.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.20.0

# Database & Storage
sqlalchemy==2.0.32
redis==5.0.8

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0

# Code Quality
black==24.8.0
flake8==7.1.1
mypy==1.11.2

# Utilities
python-json-logger==2.0.7
pyyaml==6.0.2
rich==13.7.1

# Jupyter (for notebooks)
jupyter==1.0.0
ipykernel==6.29.5
